# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PortalHatal Balmas API",
        "version": "2"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "מביא רשימת חדשות",
                "description": "מביא רשימת חדשות עם אפשרות לסינון לפי קטגוריה, חיפוש ודפדוף.",
                "operationId": "getPosts",
                "parameters": [
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "סינון לפי מאפיין הקטגוריה",
                        "required": false,
                        "schema": {
                            "type": "int",
                            "example": "1"
                        }
                    },
                    {
                        "name": "query",
                        "in": "query",
                        "description": "מילת חיפוש בכותרת, תיאור או תוכן",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": ""
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "כמות פריטים בעמוד",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "מספר עמוד לתצוגה",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "הפעולה בוצעה בהצלחה"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "יוצר פוסט חדש",
                "description": "יוצר פוסט חדש עם כותרת, תיאור, תוכן, קטגוריה ותמונה.",
                "operationId": "createPost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "content",
                                    "category_id",
                                    "image"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "כותרת הפוסט"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "תיאור קצר של הפוסט"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "{blah:{dfgfdgfd}"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "description": "תמונה מצורפת לפוסט",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "הפוסט נוצר בהצלחה"
                    },
                    "400": {
                        "description": "בקשה לא תקינה"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            }
        },
        "/api/posts/{uuid}": {
            "get": {
                "tags": [
                    "Post"
                ],
                "summary": "מביא פוסט לפי UUID",
                "description": "מביא פוסט בודד לפי מזהה ייחודי (UUID).",
                "operationId": "getPostByUUid",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID של הפוסט",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "4a206b4-99d6-4692-915c-4935766e0420"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "הפעולה בוצעה בהצלחה"
                    },
                    "400": {
                        "description": "בקשה לא תקינה"
                    },
                    "404": {
                        "description": "הפוסט לא נמצא"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            },
            "post": {
                "tags": [
                    "Post"
                ],
                "summary": "מעדכן פוסט ",
                "description": "מעדכן פוסט .",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID של הפוסט",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "4a206b4-99d6-4692-915c-4935766e0420"
                        }
                    }
                ],
                "requestBody": {
                    "required": false,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "כותרת הפוסט"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "תיאור קצר של הפוסט"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "{blah:{dfgfdgfd}"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "image": {
                                        "description": "תמונה מצורפת לפוסט",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "הפוסט נוצר בהצלחה"
                    },
                    "400": {
                        "description": "בקשה לא תקינה"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Post"
                ],
                "summary": "מוחק פוסט לפי UUID",
                "description": "מוחק פוסט בודד לפי מזהה ייחודי (UUID).",
                "operationId": "deletePostByUUid",
                "parameters": [
                    {
                        "name": "uuid",
                        "in": "path",
                        "description": "UUID של הפוסט",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid",
                            "example": "4a206b4-99d6-4692-915c-4935766e0420"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "הפעולה בוצעה בהצלחה"
                    },
                    "404": {
                        "description": "הפוסט לא נמצא"
                    },
                    "500": {
                        "description": "שגיאה בשרת"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}